+++ packages/jssdk/src/index.ts
@@ -20,6 +20,7 @@ export * from './types/catalog'
 export * from './types/bizproc/activity'
 export * from './types/bizproc'
 export * from './types/event'
+export * from './types/logger'
 
 export * from './types/b24-helper'
 export * from './types/pull'