+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * feat(RestrictionManager): new restrictions
 * feat(PlacementManager): use the property name `placement` instead of `title`
 * feat(apiVersion): support api 3
+* feat(Logger): add new logger system
 
 ## [2.1.0](https://github.com/bitrix24/b24jssdk/compare/v0.5.1...v1.1.0) (2025-12-xx)
 ## [1.1.0](https://github.com/bitrix24/b24jssdk/compare/v0.5.1...v1.1.0) (2025-12-xx)